+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Do not add Shipping tax to items
+
 ## [0.0.34] - 2021-03-11
 
 ### Fixed